# perf: remove unnecessary deepcopies and add some lru_caches