--- a/src/prettytable/prettytable.py
+++ b/src/prettytable/prettytable.py
@@ -33,7 +33,6 @@
 
 from __future__ import annotations
 
-from functools import lru_cache
 import io
 import re
 import warnings