--- a/src/prettytable/prettytable.py
+++ b/src/prettytable/prettytable.py
@@ -2705,7 +2705,7 @@ class PrettyTable:
 ##############################
 
 
-@functools.lru_cache
+@lru_cache
 def _str_block_width(val: str) -> int:
     import wcwidth  # type: ignore[import-untyped]