which led to some improvements in test performance

from
```
  lint: OK (2.01=setup[0.03]+cmd[1.98] seconds)
  mypy: OK (2.15=setup[2.02]+cmd[0.13] seconds)
  pypy3: SKIP (0.02 seconds)
  py314: SKIP (0.01 seconds)
  py313: OK (2.92=setup[1.78]+cmd[1.15] seconds)
  py312: OK (2.88=setup[1.70]+cmd[1.18] seconds)
  py311: SKIP (0.01 seconds)
  py310: SKIP (0.01 seconds)
  py39: OK (4.96=setup[3.70]+cmd[1.26] seconds)
  congratulations :) (15.02 seconds)
```
to:
```
  lint: OK (2.06=setup[0.02]+cmd[2.04] seconds)
  mypy: OK (2.16=setup[2.04]+cmd[0.11] seconds)
  pypy3: SKIP (0.01 seconds)
  py314: SKIP (0.01 seconds)
  py313: OK (2.91=setup[1.85]+cmd[1.07] seconds)
  py312: OK (2.78=setup[1.73]+cmd[1.05] seconds)
  py311: SKIP (0.01 seconds)
  py310: SKIP (0.01 seconds)
  py39: OK (4.74=setup[3.60]+cmd[1.14] seconds)
  congratulations :) (14.72 seconds)
```
which are around 7%

in an example with 25.000 rows:
```python
from prettytable import PrettyTable
import random

def add_rows(table):
    table.add_row(
        ["Adelaide", 1295, random.randint(0, 9999999), random.randint(0, 9999)]
    )
    table.add_row(
        ["Brisbane", 5905, random.randint(0, 9999999), random.randint(0, 9999)]
    )
    table.add_row(["Darwin", 112, random.randint(0, 9999999), random.randint(0, 9999)])
    table.add_row(["Hobart", 1357, random.randint(0, 9999999), random.randint(0, 9999)])
    table.add_row(["Sydney", 2058, random.randint(0, 9999999), random.randint(0, 9999)])
    table.add_row(
        ["Melbourne", 1566, random.randint(0, 9999999), random.randint(0, 9999)]
    )
    table.add_row(["Perth", 5386, random.randint(0, 9999999), random.randint(0, 9999)])

def create_table():
    table = PrettyTable(
        field_names=["City name", "Area", "Population", "Annual Rainfall"],
    )
    for i in range(5000):
        add_rows(table)
    return table

def get_table():
    t = create_table()
    return t

def main():
    t = get_table()
    s = t.get_string()

if __name__ == "__main__":
    main()

```

it improved on my machine from 1.12secs to 328msecs:

from
```
~/privat/repos/prettytable performance *1 ?2 ❯ venv/bin/python -m timeit "import test; test.main()"
1 loop, best of 5: 1.12 sec per loop
~/privat/repos/prettytable performance *1 ?2 ❯ venv/bin/python -m timeit "import test; test.main()"
1 loop, best of 5: 1.12 sec per loop
~/privat/repos/prettytable performance *1 ?2 ❯ venv/bin/python -m timeit "import test; test.main()"
1 loop, best of 5: 1.1 sec per loop
```
to
```
~/privat/repos/prettytable performance !1 ?2 ❯ venv/bin/python -m timeit "import test; test.main()"
1 loop, best of 5: 328 msec per loop
~/privat/repos/prettytable performance !1 ?2 ❯ venv/bin/python -m timeit "import test; test.main()"
1 loop, best of 5: 326 msec per loop
~/privat/repos/prettytable performance !1 ?2 ❯ venv/bin/python -m timeit "import test; test.main()"
1 loop, best of 5: 328 msec per loop
```

Those improvements are > 300%

I can't find any reason for deepcopies here. The results  of the internal `_get_rows()` function are not modified anywhere in the code. The `copy()` and the `PrettyTable()[start:end]`  functionality is not touched.